crypto-lakehouse/requirements.txt
kafka-python-ng
websockets
asyncio